Worked on Goal Class
Added a GoalBase class. Made Goal class a template inherit from GoalBase, added a function pointer for IsValid() check. Goal is a template so that any kind of data can be added to the Goal. This can be used by the IsValid() function,  or anything with access to the goal. Both template and init() have defaults for ease of use.